# specs/001-homelab-dashboard/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: HomeDash API
  version: 0.1.0
  description: |
    API contracts for HomeDash.

    Notes:
    - Cookie-based sessions are used for authentication.
    - CSRF protection is required for state-changing requests.
servers:
  - url: http://localhost:3000
...
paths:
...
  /api/admin/shell:
    get:
      tags: [admin]
      summary: Get global shell settings
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Shell settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShellSettings'
        '403':
          description: Forbidden
    put:
      tags: [admin]
      summary: Update global shell settings
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShellSettingsUpdateRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShellSettings'
  /api/admin/assets/logo:
    post:
      tags: [admin]
      summary: Upload logo (also used as favicon source)
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: Uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Asset'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: homedash_session
  parameters:
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
  schemas:
...
    Clock:
      type: object
      required: [timezone, label, isHome]
      properties:
        timezone:
          type: string
        label:
          type: string
        isHome:
          type: boolean
...
    ShellSettings:
      type: object
      required: [titleText, headerHeightPx, clockStripEnabled]
      properties:
        titleText:
          type: string
        titleFont:
          type: string
          nullable: true
        titleFontSizePx:
          type: integer
          nullable: true
        headerHeightPx:
          type: integer
        clockStripEnabled:
          type: boolean
        homeTimezone:
          type: string
          nullable: true
        timezones:
          type: array
          items:
            $ref: '#/components/schemas/Clock'
        footerText:
          type: string
          nullable: true
        repoUrl:
          type: string
          nullable: true
        unauthWebDashboardId:
          type: string
          nullable: true
        unauthMobileDashboardId:
          type: string
          nullable: true
    ShellSettingsUpdateRequest:
      allOf:
        - $ref: '#/components/schemas/ShellSettings'
    Asset:
      type: object
      required: [id, kind, url, contentType, byteSize]
      properties:
        id:
          type: string
        kind:
          type: string
          enum: [logo, background, icon]
        url:
          type: string
        contentType:
          type: string
        byteSize:
          type: integer
...